move methods for filternaming back fromFilterUtil to GeocacheFilter
to avoid dependency to gui in GeocacheFilter
