+++ main/src/main/java/cgeo/geocaching/filters/gui/GeocacheFilterActivity.java
@@ -163,7 +163,7 @@ public class GeocacheFilterActivity extends AbstractActionBarActivity {
 
         //handling of "save" button
         binding.filterStorageSave.setOnClickListener(v -> {
-            final String filterName = FilterUtils.getPurifiedFilterName(binding.filterStorageName.getText().toString());
+            final String filterName = GeocacheFilter.getPurifiedFilterName(binding.filterStorageName.getText().toString());
             SimpleDialog.of(this).setTitle(R.string.cache_filter_storage_save_title)
                     .input(new SimpleDialog.InputOptions().setInitialValue(filterName), newName -> {
                         final GeocacheFilter filter = getFilterFromView();
@@ -401,7 +401,7 @@ public class GeocacheFilterActivity extends AbstractActionBarActivity {
         }
 
         return GeocacheFilter.create(
-                FilterUtils.getPurifiedFilterName(binding.filterStorageName.getText().toString()),
+                GeocacheFilter.getPurifiedFilterName(binding.filterStorageName.getText().toString()),
                 binding.filterBasicAdvanced.isChecked(),
                 this.includeInconclusiveFilterCheckbox.isChecked(),
                 filter);
